refactor: auto-split test_simulator.py
